german_credit_data_biased_training
A CSV file is a text file that has a specific format which allows data to be saved in a table structured format.
Note: This CSV file is a publicly available file that has sample, fictitious data in it for training purposes. It’s not intended for any other purposes. It’s actually available in an IBM GitHub repository, too!